API change with arithmetic functions now returning {:ok, value} or {:error, reason} "bang" methods are provided that exhibit the previous behaviour.  This change is to better conform with idiomatic Exlir/Erlang behaviour.
Also document the way to configure the exchange rate service manually which
makes it possible to set the application start order and therefore let callback
modules use Ecto or other apps.

Some refactoring to be more idiomatic on pipeline formatting.
Changed the retrieval interval to 300_000 (5 minutes) which is what
the documentation says.

Closes #15, #16