Refactors Money.{Arithmetic, Conversion} functions into Money.
Money.Financial functions are kept separately and need to be invoked on that module, not on Money - this part is a breaking change.

In addition:
a. Improve the readme
b. Add the changelog to the generated online docs
c. Improve test coverage
d. Fix a bug on Money.new!(currency, decimal) and Money.new!(decimal, currency)