Separate exchange rate concerns 1. Introduce ExchangeRates.Cache as the single responsibility for cache update and access 2. Move all retriever API functions into the Retriever module with the same access API as the Cache 3. Add etag support to the OpenExchangeRates retriever
With this commit, the functions latest_rates/1 and
historic_rates/1 are implemented in the module ExchangeRates.Cache
and ExchangeRates.Retriever.  The Cache version returns cached
results, the Retriever one returns API results.

The public API in ExchangeRates still needs to be modified
to return cached results if available, API results if not,
and perhaps an argument to force it either way.