Money.Subscription.new/3 replaces Money.Subscription.new/1 Money.Subscription.new now takes a plan and effective date and some options.
The start of the current interval for a subscription is
now calculated from the plan history.  This makes determining
the date of the current period simpler at the expense of
some additional computation.  The current algorithm is not
optimal since it recursively increments next_interval_date
until today is in the range current_interval_starts..
next_interval_date.