Adds a migration to add aggregate functions for money_with_currency Postgres requires a customer aggregation function for user defined types.
The migration adds the `sum` aggregate for the type
`money_with_currency`.

When migrated, this function allows the use of the standard
Ecto API for the `sum()` function to also operate on
`money_with_currency` columns.

Since aggregation only makes sense if the currency code
of all the selected rows is the same an exception is
raised in the case the mixed currency codes are detected.