Support Elixir 1.10 Enum.sort
Elixir 1.10 introduces `Enum.sort/2` that will call
the function `compare/2` on a module that is expected
to return :lt, :eq or :gt.

This has necessitated a swap of the functions `compare/2`
and `cmp/2` in `ex_money` to deliver the compatible results.

Also add support for upcoming Decimal releases that are
also adapting to this change over the 1.9 and 2.0 releases.